DeskClock: fix behavior with alarms when device is off
The current behavior causes an NPE and prevents the alarm
from firing. Determine if this a boot alarm and if the
alarm needs to be fired immediately. If not, proceed with
normal alarm handling.

Also, adding some checks to prevent NPEs and FCs when
handling an intent without data() and querying for an
AlarmInstance with an invalid id.

Change-Id: I0d5ee998c43207905875a0272fe70596a3877587